requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
pytz==2020.1
requests==2.24.0
sqlparse==0.4.0
urllib3==1.25.10
